Cucumber-ML-main/requirements.txt
absl-py==2.0.0
astunparse==1.6.3
atomicwrites==1.4.0
brotlipy==0.7.0
cachetools==5.3.2
contourpy==1.1.1
cycler==0.12.1
flatbuffers==23.5.26
fonttools==4.43.1
future==0.18.3
gast==0.5.4
google-auth==2.23.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.59.0
h5py==3.10.0
inflection==0.5.1
joblib==1.3.2
keras==2.14.0
kiwisolver==1.4.5
libclang==16.0.6
Markdown==3.5
matplotlib==3.8.0
ml-dtypes==0.2.0
numpy==1.26.0
oauthlib==3.2.2
opt-einsum==3.3.0
pandas==2.1.1
Pillow==10.0.1
ply==3.11
protobuf==4.24.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyls-spyder==0.4.0
pyparsing==3.1.1
PyQt5==5.15.7
PyQtWebEngine==5.15.4
pyswarms==1.3.0
python-lsp-jsonrpc==1.0.0
pywin32==305.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.3.1
scipy==1.11.3
seaborn==0.13.0
tensorboard==2.14.1
tensorboard-data-server==0.7.2
tensorflow==2.14.0
tensorflow-estimator==2.14.0
tensorflow-intel==2.14.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
threadpoolctl==3.2.0
tqdm==4.66.1
tzdata==2023.3
webencodings==0.5.1
Werkzeug==3.0.1
scikit-learn==1.3.1
joblib==1.3.2
neuralprophet